refactor: data interfaces
